add markToDate feature to calendar-date type
Add this to enable visual marking of date ranges by applying selected
styling to dates between selected date and markToDate.

This also requires me to add the `isDateBetween` utility function for
two ways of date range checking.

Note: My use case is to create 2 calendar-date components where one
marks the range based on the other's selection. It is working as
expected but not sure if it might cause any issues on other component
types.